requirements.txt
asttokens==2.4.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
Babel==2.8.0
base58==1.0.3
bcrypt==3.2.0
blinker==1.4
bottle==0.12.25
Brotli==1.0.9
cbor==1.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
cloud-init==23.1.2
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
Cython==0.29.28
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.18.0b1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.12.2
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.45.0
GeoIP==1.3.2
hkdf==0.0.3
httplib2==0.20.2
humanize==0.0.0
hyperlink==21.0.0
idna==3.3
imbalanced-learn==0.11.0
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
kafka-python==2.0.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libvirt-python==8.0.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
magic-wormhole==0.12.0
MarkupSafe==2.0.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mne==1.6.0
mnemonic==0.19
more-itertools==8.10.0
msgpack==1.0.3
mutagen==1.45.1
nest-asyncio==1.5.8
netifaces==0.11.0
norminette==3.3.51
numpy==1.26.2
oauthlib==3.2.0
packaging==21.3
pandas==2.1.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.10.0
ply==3.11
pooch==1.8.0
prettytable==3.9.0
prompt-toolkit==3.0.41
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-magic==0.4.24
python-snappy==0.5.3
PyTrie==0.4.0
pytz==2022.1
PyWavelets==1.5.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
requests==2.25.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sortedcontainers==2.1.0
sos==4.4
spake2==0.8
speedtest-cli==2.1.3
ssh-import-id==5.11
stack-data==0.6.3
supervisor==4.2.1
systemd-python==234
terminator==2.1.1
texttable==1.6.4
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.57.0
traitlets==5.13.0
Twisted==22.1.0
txaio==21.2.1
txtorcon==20.0.0
tzdata==2023.3
u-msgpack-python==2.3.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
urllib3==1.26.5
vboxapi==1.0
virtualenv==20.24.3
wadllib==1.3.6
wcwidth==0.2.12
websocket-client==1.2.3
websockets==9.1
wsaccel==0.6.3
xdg==5
xgboost==2.0.2
xkit==0.0.0
yt-dlp==2022.4.8
zipp==1.0.0
zope.interface==5.4.0
